Revert "Merge pull request #320 from asottile/new_class_super_v2"
This reverts commit b3f8c7b2e9ccb06c6028d5dd90c8f5490a49e95b, reversing
changes made to 41308214566af7c5589ebc7e8a760522c82dc3bc.
