+++ tests/features/yield_from_test.py
@@ -1,7 +1,11 @@
+import ast
+
 import pytest
 
 from pyupgrade._data import Settings
 from pyupgrade._main import _fix_plugins
+from pyupgrade._plugins.legacy import _fields_same
+from pyupgrade._plugins.legacy import _targets_same
 
 
 @pytest.mark.parametrize(
@@ -211,3 +215,18 @@ def test_fix_yield_from(s, expected):
 )
 def test_fix_yield_from_noop(s):
     assert _fix_plugins(s, settings=Settings(min_version=(3,))) == s
+
+
+def test_targets_same():
+    assert _targets_same(ast.parse('global a, b'), ast.parse('global a, b'))
+    assert not _targets_same(ast.parse('global a'), ast.parse('global b'))
+
+
+def _get_body(expr):
+    body = ast.parse(expr).body[0]
+    assert isinstance(body, ast.Expr)
+    return body.value
+
+
+def test_fields_same():
+    assert not _fields_same(_get_body('x'), _get_body('1'))
